requirements.txt
boto3==1.42.55
discord.py==2.6.4
httpx==0.28.1
playwright==1.58.0
python-dotenv==1.2.1
